fix: 메인 - 기정의한 모델 제거하고 로비 엔티티로 정의 (#18)
