+++ Meme/Meme/Sources/Wiki/Data/DTO/LobbyResponse.swift
@@ -8,18 +8,5 @@
 import Foundation
 
 struct LobbyResponse: Decodable {
-    let sections: [SectionResponse]
     
-    func toEntity() -> Lobby {
-        Lobby(sections: sections.map { $0.toEntity() })
-    }
-    
-    struct SectionResponse: Decodable {
-        let type: String
-        let title: String
-        
-        func toEntity() -> Lobby.Section {
-            Lobby.Section(type: type, title: title)
-        }
-    }
 }
