+++ Meme/Meme/Sources/Wiki/Presentation/Main/View/Cell/MemeMainTopRatedCell.swift
@@ -17,7 +17,7 @@ final class MemeMainTopRatedCell: UICollectionViewCell {
         }
     }
     
-    func configureCell(with item: MemeMain.Model.Item) {
+    func configureCell(with item: Lobby.Item) {
         contentView.layer.cornerRadius = 12
         contentView.layer.masksToBounds = true
         contentView.backgroundColor = .systemPink.withAlphaComponent(0.3)
